+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 5.10.3.0 - 2026-06-XX (NINA 3.3)
 * Added a Moon Avoidance Helper to assist with understanding avoidance behavior and selecting parameters
+* Added new image grader preferences: auto reject levels for HFR, FWHM, and Eccentricity.
 * Fix for new preview report timing issues
 * Added chart to preview report as well as HTML output